Update and Delete dialog
Added a new dialog layout
to update and delete added expenses. Known issues is that when picking a time in the main add expense the clock numbers are white which is confusing. As well as in the updatedialog the update category isn't in the proper list. Need to fix all later.